remove unused rpc impl
