+++ crates/client/src/lib.rs
@@ -10,7 +10,7 @@ use cfxkey as keylib;
 pub mod accounts;
 pub mod common;
 mod node_types;
-pub mod rpc;
 pub use cfx_config as configuration;
 pub use node_types::{archive, full, light};
+pub mod rpc_starter;
 pub mod state_dump;
