--- crates/client/src/rpc/errors.rs
@@ -1,3 +1,0 @@
-pub use cfx_rpc_utils::error::{
-    error_codes, error_codes as codes, errors::*, jsonrpc_error_helpers::*,
-};
