--- crates/client/src/rpc/helpers/mod.rs
@@ -1,12 +1,0 @@
-// Copyright 2020 Conflux Foundation. All rights reserved.
-// Conflux is free software and distributed under GNU General Public License.
-// See http://www.gnu.org/licenses/
-
-use cfx_rpc_cfx_impl::helpers::block_provider;
-mod subscribers;
-
-pub use block_provider::{build_block, build_header};
-pub use cfx_rpc_eth_impl::helpers::{
-    EpochQueue, MAX_FEE_HISTORY_CACHE_BLOCK_COUNT,
-};
-pub use subscribers::{Id as SubscriberId, Subscribers};
