--- crates/client/src/rpc/impls/cfx/mod.rs
@@ -1,10 +1,0 @@
-pub mod cfx_filter;
-pub mod cfx_handler;
-pub mod common;
-pub mod light;
-pub mod pool;
-pub mod pubsub;
-mod trace;
-
-pub use cfx_handler::{CfxHandler, LocalRpcImpl, RpcImpl, TestRpcImpl};
-pub use trace::TraceHandler;
