--- crates/client/src/rpc/impls/mod.rs
@@ -1,9 +1,0 @@
-// Copyright 2019 Conflux Foundation. All rights reserved.
-// Conflux is free software and distributed under GNU General Public License.
-// See http://www.gnu.org/licenses/
-
-pub mod cfx;
-pub mod pos;
-
-pub use cfx::{cfx_filter, common, light, pool, pubsub};
-pub use cfx_rpc_cfx_types::{FeeHistoryCacheEntry, RpcImplConfiguration};
