--- crates/client/src/rpc/traits/cfx_space/mod.rs
@@ -1,8 +1,0 @@
-pub mod cfx;
-pub mod cfx_filter;
-pub mod debug;
-pub mod pool;
-pub mod pos;
-pub mod pubsub;
-pub mod test;
-pub mod trace;
