--- crates/client/src/rpc/traits/mod.rs
@@ -1,7 +1,0 @@
-// Copyright 2019 Conflux Foundation. All rights reserved.
-// Conflux is free software and distributed under GNU General Public License.
-// See http://www.gnu.org/licenses/
-
-pub use cfx_space::{cfx, cfx_filter, debug, pool, pos, pubsub, test, trace};
-
-pub mod cfx_space;
