--- crates/client/src/rpc/types/cfx.rs
@@ -1,16 +1,0 @@
-pub use cfx_rpc_cfx_types::{
-    access_list::*,
-    account::{self, Account},
-    address,
-    address::{
-        check_rpc_address_network, check_two_rpc_address_network_match,
-        RcpAddressNetworkInconsistent, RpcAddress, UnexpectedRpcAddressNetwork,
-    },
-    blame_info, block, consensus_graph_states, epoch_number, filter, log,
-    pos_economics, pubsub, receipt, reward_info, sponsor_info,
-    stat_on_gas_load, status, storage_collateral_info, sync_graph_states,
-    token_supply_info, transaction, transaction_request, tx_pool,
-    vote_params_info, CfxFeeHistory,
-};
-pub use sponsor_info::SponsorInfo;
-pub use tx_pool::*;
