--- crates/client/src/rpc/types/mod.rs
@@ -1,50 +1,0 @@
-// Copyright 2019 Conflux Foundation. All rights reserved.
-// Conflux is free software and distributed under GNU General Public License.
-// See http://www.gnu.org/licenses/
-
-pub mod cfx;
-
-pub use self::cfx::{
-    address::{self, check_two_rpc_address_network_match, RpcAddress},
-    blame_info::BlameInfo,
-    block::{Block, BlockTransactions, Header},
-    consensus_graph_states::{
-        ConsensusGraphBlockExecutionState, ConsensusGraphBlockState,
-        ConsensusGraphStates,
-    },
-    epoch_number::{BlockHashOrEpochNumber, EpochNumber},
-    filter::{CfxFilterChanges, CfxFilterLog, CfxRpcLogFilter, RevertTo},
-    log::Log,
-    pos_economics::PoSEconomics,
-    pubsub,
-    receipt::Receipt,
-    reward_info::RewardInfo,
-    stat_on_gas_load::StatOnGasLoad,
-    status::Status,
-    storage_collateral_info::StorageCollateralInfo,
-    sync_graph_states::{SyncGraphBlockState, SyncGraphStates},
-    token_supply_info::TokenSupplyInfo,
-    transaction::{PackedOrExecuted, Transaction},
-    transaction_request::{
-        self, CheckBalanceAgainstTransactionResponse,
-        EstimateGasAndCollateralResponse, TransactionRequest,
-        DEFAULT_CFX_GAS_CALL_REQUEST,
-    },
-    tx_pool::{
-        AccountPendingInfo, AccountPendingTransactions,
-        TxPoolPendingNonceRange, TxPoolStatus, TxWithPoolInfo,
-    },
-    vote_params_info::VoteParamsInfo,
-    Account, CfxFeeHistory, SponsorInfo,
-};
-pub use cfx_rpc_cfx_types::{
-    pos,
-    provenance::Origin,
-    trace::{
-        Action, LocalizedBlockTrace, LocalizedTrace, LocalizedTransactionTrace,
-    },
-    trace_filter::TraceFilter,
-};
-pub use cfx_rpc_primitives::{Bytes, Index, U64};
-
-pub use cfx_rpc_eth_types::{FeeHistory, WrapTransaction};
